bootloader: fix two nasty bugs
We had 2 very nasty bugs on bootloader. One was that SP was set too high, so stack would underflow eventually.
We also had "ivector" at a "random" address, so when we shadowed sketch and set ivector, we were actually overwriting
parts of that sketch. So both bootloader and sketch have a "boots" section now which starts at 0x1000 and includes ivector
and sketch entry point.